Remove urlz package; inline its one live function into cli
The libsq/core/urlz package had a single non-test caller: StripSchemeAndUser,
used by shell-completion location suggestions. Its other six exported
functions (QueryParamKeys, RenameQueryParamKey, StripQuery, StripPath,
StripUser, StripScheme) had no callers.

Move StripSchemeAndUser to package cli as the unexported helper
stripSchemeAndUser, next to its sole caller, and delete the urlz package.
The helper takes a url.URL, so cli (which already imports net/url) is a
better home than a string-utilities package. Its test moves along with it.
